Updating the download link to reference the Piriform destination that automatically downloads the latest version.
This was changed as the other link will not download the 'slim' version (One without bloatware) - Unforunately
there was no link to the the slim version that was versioned. After checking with Rob Reynolds (Maintainer of Chocolatey.org
he stated that it was better to have a non-versioned link and no bloatware so the change was made.